AE 67  Major: Added tracking of file type so it doesn't get lost
Previously, file category of normal vs tumour would get lost thorugh the multiple steps. Added the identity of the file to a variable which tracks throughout the various processes so there are no loss an no confusion.

This is important to make sure we don't get confused with tumor vs normal calculations.