AE69 Call heteroplasmy bug fix
Fixed a bug arising in final perl call heteroplasmy script. This was a result of changes made for the mitocaller2vcf process. However, solution was to pipe out both a tsv file and a tsv.gz file out of the call_mitocaller proces.

Thanks to  @rhughwhite  for letting me know about the bug